requirements.txt
astroid==2.5
cffi==1.14.5
click==7.1.2
cryptography==3.4.6
Flask==1.1.2
Flask-Cors==3.0.10
isort==5.7.0
itsdangerous==1.1.0
Jinja2==2.11.3
lazy-object-proxy==1.5.2
MarkupSafe==1.1.1
mccabe==0.6.1
pycparser==2.20
PyJWT==2.0.1
pylint==2.7.1
python-dotenv==0.15.0
six==1.15.0
toml==0.10.2
Werkzeug==1.0.1
wrapt==1.12.1
